AiAnalysis progress, only pdf and specific image formats, add try catch and loggin to some methods, remove and add user credits

+++ API/DTOs/JobCandidatesTableDataDto.cs
@@ -18,10 +18,10 @@ namespace API.DTOs
         public DateTime ApplicationDate { get; set; }
         public DateTime? MeetingDateTime { get; set; }
         public bool? IsOnlineMeeting { get; set; }
-        public int? AIMatchingResult { get; set; }
-        public int? AIMatchingExperience { get; set; }
-        public int? AIMatchingSkills { get; set; }
-        public int? AIMatchingEducationLevel { get; set; }
+        public double? AIMatchingResult { get; set; }
+        public double? AIMatchingExperience { get; set; }
+        public double? AIMatchingSkills { get; set; }
+        public double? AIMatchingEducationLevel { get; set; }
         public string AIMatchingDescription { get; set; }
         public int? ConversationId { get; set; }
     }
