+++ API/Entities/Applications/UserApplication.cs
@@ -46,10 +46,10 @@ namespace API.Entities.Applications
         public ICollection<UserApplicationEducation> Educations { get; set; }
         public ICollection<UserApplicationPreviousCompanies> PreviousCompanies { get; set; }
 
-        public int? AIMatchingResult { get; set; }
-        public int? AIMatchingExperience { get; set; }
-        public int? AIMatchingSkills { get; set; }
-        public int? AIMatchingEducationLevel { get; set; }
+        public double? AIMatchingResult { get; set; }
+        public double? AIMatchingExperience { get; set; }
+        public double? AIMatchingSkills { get; set; }
+        public double? AIMatchingEducationLevel { get; set; }
         public string AIMatchingDescription { get; set; }
     }
